Extract live send run activation
